+++ docs/source/conf.py
@@ -13,19 +13,21 @@
 import os.path
 import sys
 
-sys.path[:0] = [os.path.dirname(os.path.abspath(__file__)) + "/../../"]
+sys.path.insert(0, os.path.abspath("."))
+sys.path.insert(0, os.path.abspath("../"))
+sys.path.insert(0, os.path.abspath("../../"))
+
 import dscript
 
 
 # -- Project information -----------------------------------------------------
 
 project = "D-SCRIPT"
-copyright = "2020, Samuel Sledzieski, Rohit Singh"
+copyright = "2021, Samuel Sledzieski, Rohit Singh"
 author = "Samuel Sledzieski, Rohit Singh"
 
 # The full version, including alpha/beta/rc tags
-release = "v1.0-beta"
-
+release = "v2.0"
 master_doc = "index"