+++ dscript/__init__.py
@@ -1,12 +1,23 @@
-__version__ = "0.1.9"
-__citation__ = """Sledzieski, Singh, Cowen, Berger. Sequence-based prediction of protein-protein interactions: a structure-aware interpretable deep learning model. Cell Systems, 2021."""
-from . import alphabets, commands, fasta, language_model, models, pretrained
+__version__ = "0.2.0"
+__citation__ = """Sledzieski, Singh, Cowen, Berger. "D-SCRIPT translates genome to phenome with sequence-based, structure-aware, genome-scale predictions of protein-protein interactions." Cell Systems 12, no. 10 (2021): 969-982.
+
+Devkota, Singh, Sledzieski, Berger, Cowen, Topsy-Turvy: integrating a global view into sequence-based PPI prediction, Bioinformatics, In Press."""
+from . import (
+    alphabets,
+    commands,
+    fasta,
+    glider,
+    language_model,
+    models,
+    pretrained,
+)
 
 __all__ = [
     "models",
     "commands",
     "alphabets",
     "fasta",
+    "glider",
     "language_model",
     "pretrained",
     "utils",