+++ dscript/fasta.py
@@ -1,51 +1,16 @@
-from __future__ import print_function, division
-
-
-def parse_stream(f, comment=b"#"):
-
-    name = None
-    sequence = []
-    for line in f:
-        if line.startswith(comment):
-            continue
-        line = line.strip()
-        if line.startswith(b">"):
-            if name is not None:
-                yield name, b"".join(sequence)
-            name = line[1:]
-            sequence = []
-        else:
-            sequence.append(line.upper())
-    if name is not None:
-        yield name, b"".join(sequence)
+import os
+from Bio import SeqIO
+from Bio.Seq import Seq
+from Bio.SeqRecord import SeqRecord
 
 
 def parse(f, comment="#"):
-    starter = ">"
-    empty = ""
-    if "b" in f.mode:
-        comment = b"#"
-        starter = b">"
-        empty = b""
     names = []
     sequences = []
-    name = None
-    sequence = []
-    for line in f:
-        if line.startswith(comment):
-            continue
-        line = line.strip()
-        if line.startswith(starter):
-            if name is not None:
-                names.append(name)
-                sequences.append(empty.join(sequence))
-            name = line[1:]
-            sequence = []
-        else:
-            sequence.append(line.upper())
-    if name is not None:
-        names.append(name)
-        sequences.append(empty.join(sequence))
+
+    for record in SeqIO.parse(f, "fasta"):
+        names.append(record.name)
+        sequences.append(str(record.seq))
 
     return names, sequences
 
@@ -56,16 +21,18 @@ def parse_directory(directory, extension=".seq"):
 
     for seqPath in os.listdir(directory):
         if seqPath.endswith(extension):
-            n, s = parse(open(f"{directory}/{seqPath}", "rb"))
+            n, s = parse(f"{directory}/{seqPath}", "rb")
             names.append(n[0].decode("utf-8").strip())
             sequences.append(s[0].decode("utf-8").strip())
     return names, sequences
 
 
 def write(nam, seq, f):
-    for n, s in zip(nam, seq):
-        f.write(">{}\n".format(n))
-        f.write("{}\n".format(s))
+    records = [
+        SeqRecord(Seq(s), id=n, name=n, description="")
+        for n, s in zip(nam, seq)
+    ]
+    SeqIO.write(records, f, "fasta")
 
 
 def count_bins(array, bins):